Moved GeneratedOrderKey
