+++ service/src/main/java/uk/gov/hmcts/reform/fpl/enums/ccd/casefields/GeneratedOrderKey.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.fpl.enums;
+package uk.gov.hmcts.reform.fpl.enums.ccd.casefields;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum GeneratedOrderKey {
+public enum GeneratedOrderKey implements CaseField {
     ORDER_TYPE_AND_DOCUMENT("orderTypeAndDocument"),
     ORDER("order"),
     JUDGE_AND_LEGAL_ADVISOR("judgeAndLegalAdvisor"),
